+++ pinyin.go
@@ -7,7 +7,7 @@ import (
 
 // Meta
 const (
-	Version   = "0.10.0"
+	Version   = "0.11.0"
 	Author    = "mozillazg, 闲耘"
 	License   = "MIT"
 	Copyright = "Copyright (c) 2016 mozillazg, 闲耘"
@@ -17,12 +17,14 @@ const (
 const (
 	Normal      = 0 // 普通风格，不带声调（默认风格）。如： zhong guo
 	Tone        = 1 // 声调风格1，拼音声调在韵母第一个字母上。如： zhōng guó
-	Tone2       = 2 // 声调风格2，即拼音声调在各个拼音之后，用数字 [0-4] 进行表示。如： zho1ng guo2
+	Tone2       = 2 // 声调风格2，即拼音声调在各个韵母之后，用数字 [1-4] 进行表示。如： zho1ng guo2
+	Tone3       = 8 // 声调风格3，即拼音声调在各个拼音之后，用数字 [1-4] 进行表示。如： zhong1 guo2
 	Initials    = 3 // 声母风格，只返回各个拼音的声母部分。如： zh g
 	FirstLetter = 4 // 首字母风格，只返回拼音的首字母部分。如： z g
-	Finals      = 5 // 韵母风格1，只返回各个拼音的韵母部分，不带声调。如： ong uo
-	FinalsTone  = 6 // 韵母风格2，带声调，声调在韵母第一个字母上。如： ōng uó
-	FinalsTone2 = 7 // 韵母风格2，带声调，声调在各个拼音之后，用数字 [0-4] 进行表示。如： o1ng uo2
+	Finals      = 5 // 韵母风格，只返回各个拼音的韵母部分，不带声调。如： ong uo
+	FinalsTone  = 6 // 韵母风格1，带声调，声调在韵母第一个字母上。如： ōng uó
+	FinalsTone2 = 7 // 韵母风格2，带声调，声调在各个韵母之后，用数字 [1-4] 进行表示。如： o1ng uo2
+	FinalsTone3 = 9 // 韵母风格3，带声调，声调在各个拼音之后，用数字 [1-4] 进行表示。如： ong1 uo2
 )
 
 // 拼音风格(兼容之前的版本)
@@ -56,7 +58,10 @@ var rePhoneticSymbolSource = func(m map[string]string) string {
 var rePhoneticSymbol = regexp.MustCompile("[" + rePhoneticSymbolSource + "]")
 
 // 匹配使用数字标识声调的字符的正则表达式
-var reTone2 = regexp.MustCompile("([aeoiuvnm])([0-4])$")
+var reTone2 = regexp.MustCompile("([aeoiuvnm])([1-4])$")
+
+// 匹配 Tone2 中标识韵母声调的正则表达式
+var reTone3 = regexp.MustCompile("^([a-z]+)([1-4])([a-z]*)$")
 
 // Args 配置信息
 type Args struct {
@@ -159,7 +164,7 @@ func toFixed(p string, a Args) string {
 		case Normal, FirstLetter, Finals:
 			// 去掉声调: a1 -> a
 			m = reTone2.ReplaceAllString(symbol, "$1")
-		case Tone2, FinalsTone2:
+		case Tone2, FinalsTone2, Tone3, FinalsTone3:
 			// 返回使用数字标识声调的字符
 			m = symbol
 		default:
@@ -169,11 +174,16 @@ func toFixed(p string, a Args) string {
 	})
 
 	switch a.Style {
+	// 将声调移动到最后
+	case Tone3, FinalsTone3:
+		py = reTone3.ReplaceAllString(py, "$1$3$2")
+	}
+	switch a.Style {
 	// 首字母
 	case FirstLetter:
 		py = py[:1]
 	// 韵母
-	case Finals, FinalsTone, FinalsTone2:
+	case Finals, FinalsTone, FinalsTone2, FinalsTone3:
 		// 转换为 []rune unicode 编码用于获取第一个拼音字符
 		// 因为 string 是 utf-8 编码不方便获取第一个拼音字符
 		rs := []rune(origP)